requirements.txt
# ── Core Framework ───────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0

# ── LangChain + LangGraph ───────────────────────────────────────
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# ── Vector DB ────────────────────────────────────────────────────
qdrant-client>=1.12.0

# ── HTTP Client ──────────────────────────────────────────────────
httpx>=0.27.0

# ── Dev ──────────────────────────────────────────────────────────
python-dotenv>=1.0.0
# ── Web Tools & MCP ──────────────────────────────────────────────
ddgs>=9.13.0
beautifulsoup4
langchain-community
fastmcp>=2.0.0
langchain-mcp-adapters>=0.1.0

# ── PPTX Generator ───────────────────────────────────────────────
python-pptx>=1.0.0

# ── File Upload (Audio ASR) ──────────────────────────────────────
python-multipart>=0.0.12
lxml>=5.0.0

# ── Document Parser ──────────────────────────────────────────────
PyMuPDF>=1.24.0
pandas>=2.2.0
python-docx>=1.1.0
aiofiles>=23.2.0
tabulate>=0.9.0          # required by pandas DataFrame.to_markdown()
matplotlib>=3.8.0
seaborn>=0.13.0